+++ projects/packages/videopress/src/client/block-editor/blocks/video/edit.tsx
@@ -27,12 +27,12 @@ import ColorPanel from './components/color-panel';
 import DetailsPanel from './components/details-panel';
 import { VideoPressIcon } from './components/icons';
 import PlaybackPanel from './components/playback-panel';
+import Player from './components/player';
 import PosterImageBlockControl from './components/poster-image-block-control';
 import PosterPanel from './components/poster-panel';
 import PrivacyAndRatingPanel from './components/privacy-and-rating-panel';
 import ReplaceControl from './components/replace-control';
 import TracksControl from './components/tracks-control';
-import VideoPressPlayer from './components/videopress-player';
 import VideoPressUploader from './components/videopress-uploader';
 import { description, title } from '.';
 /**
@@ -554,7 +554,7 @@ export default function VideoPressEdit( {
 				} }
 			/>
 
-			<VideoPressPlayer
+			<Player
 				html={ html }
 				isRequestingEmbedPreview={ isRequestingEmbedPreview }
 				scripts={ scripts }
